Refactoring: Introduce `overlay/caching.ts`
