+++ lib/resolve-environment-action.js
@@ -93210,7 +93210,7 @@ var require_cacheHttpClient = __commonJS({
     exports2.getCacheEntry = getCacheEntry;
     exports2.downloadCache = downloadCache;
     exports2.reserveCache = reserveCache;
-    exports2.saveCache = saveCache4;
+    exports2.saveCache = saveCache3;
     var core14 = __importStar2(require_core());
     var http_client_1 = require_lib();
     var auth_1 = require_auth();
@@ -93387,7 +93387,7 @@ Other caches with similar key:`);
         }));
       });
     }
-    function saveCache4(cacheId, archivePath, signedUploadURL, options) {
+    function saveCache3(cacheId, archivePath, signedUploadURL, options) {
       return __awaiter2(this, void 0, void 0, function* () {
         const uploadOptions = (0, options_1.getUploadOptions)(options);
         if (uploadOptions.useAzureSdk) {
@@ -98887,8 +98887,8 @@ var require_cache5 = __commonJS({
     Object.defineProperty(exports2, "__esModule", { value: true });
     exports2.FinalizeCacheError = exports2.ReserveCacheError = exports2.ValidationError = void 0;
     exports2.isFeatureAvailable = isFeatureAvailable;
-    exports2.restoreCache = restoreCache4;
-    exports2.saveCache = saveCache4;
+    exports2.restoreCache = restoreCache3;
+    exports2.saveCache = saveCache3;
     var core14 = __importStar2(require_core());
     var path7 = __importStar2(require("path"));
     var utils = __importStar2(require_cacheUtils());
@@ -98945,7 +98945,7 @@ var require_cache5 = __commonJS({
           return !!process.env["ACTIONS_CACHE_URL"];
       }
     }
-    function restoreCache4(paths_1, primaryKey_1, restoreKeys_1, options_1) {
+    function restoreCache3(paths_1, primaryKey_1, restoreKeys_1, options_1) {
       return __awaiter2(this, arguments, void 0, function* (paths, primaryKey, restoreKeys, options, enableCrossOsArchive = false) {
         const cacheServiceVersion = (0, config_1.getCacheServiceVersion)();
         core14.debug(`Cache service version: ${cacheServiceVersion}`);
@@ -99089,7 +99089,7 @@ var require_cache5 = __commonJS({
         return void 0;
       });
     }
-    function saveCache4(paths_1, key_1, options_1) {
+    function saveCache3(paths_1, key_1, options_1) {
       return __awaiter2(this, arguments, void 0, function* (paths, key, options, enableCrossOsArchive = false) {
         const cacheServiceVersion = (0, config_1.getCacheServiceVersion)();
         core14.debug(`Cache service version: ${cacheServiceVersion}`);
@@ -104133,7 +104133,6 @@ var semver5 = __toESM(require_semver2());
 // src/overlay/index.ts
 var fs3 = __toESM(require("fs"));
 var path4 = __toESM(require("path"));
-var actionsCache = __toESM(require_cache5());
 
 // src/git-utils.ts
 var fs2 = __toESM(require("fs"));
@@ -104329,8 +104328,6 @@ var CODEQL_OVERLAY_MINIMUM_VERSION_JAVA = "2.23.8";
 var CODEQL_OVERLAY_MINIMUM_VERSION_JAVASCRIPT = "2.23.9";
 var CODEQL_OVERLAY_MINIMUM_VERSION_PYTHON = "2.23.9";
 var CODEQL_OVERLAY_MINIMUM_VERSION_RUBY = "2.23.9";
-var OVERLAY_BASE_DATABASE_MAX_UPLOAD_SIZE_MB = 7500;
-var OVERLAY_BASE_DATABASE_MAX_UPLOAD_SIZE_BYTES = OVERLAY_BASE_DATABASE_MAX_UPLOAD_SIZE_MB * 1e6;
 async function writeBaseDatabaseOidsFile(config, sourceRoot) {
   const gitFileOids = await getFileOidsUnderPath(sourceRoot);
   const gitFileOidsJson = JSON.stringify(gitFileOids);
@@ -104650,10 +104647,10 @@ var featureConfig = {
 };
 
 // src/overlay/status.ts
-var actionsCache2 = __toESM(require_cache5());
+var actionsCache = __toESM(require_cache5());
 
 // src/trap-caching.ts
-var actionsCache3 = __toESM(require_cache5());
+var actionsCache2 = __toESM(require_cache5());
 
 // src/config-utils.ts
 var OVERLAY_MINIMUM_AVAILABLE_DISK_SPACE_MB = 2e4;
