+++ src/overlay/overlay-database-mode.ts
@@ -1,0 +1,5 @@
+export enum OverlayDatabaseMode {
+  Overlay = "overlay",
+  OverlayBase = "overlay-base",
+  None = "none",
+}
